Refacotr GafanaAggragtor
Refactor to hold events in memory for the `event_aggragator_time_window`
and then send them in a burst
